Cool VL Viewer v1.26.4.13 (current stable branch):
New in this release:
Added spell a checking feature. See this message for details.
Added a confirmation/warning dialog when quitting the viewer while some notification(s) are still unacknowledged.
RestrainedLove: lifted the maximum height correction limits in @adjustheight so to deal with the new micro or macro mesh-based avatars.
For Linux builds only: fixed a crash bug caused by the newest fontconfig v2.9. See this thread for details.
For Linux: added an option (LL_REPORT) in the coll_vl_viewer wrapper script that, when enabled, makes the latter script report any success or failure of the viewer session via a dialog (Xdialog, Zenity or xmessage, whichever is present on your system) and presents the stack trace log when it was created after a crash.