Cool VL Viewer v1.26.4.14 (current stable branch):
New in this release:
Added a Text Input floater for the chat and IM input lines: you can now configure the viewer (in "Preferences" -> "Cool features" -> "User Interface") to auto-open a text input floater automatically whenever the chat or IM input lines get so full that they must be scrolled left. A small ">" button was also added on the right of the input lines to manually open the text input floater. The advantage of the latter is that it is a multi-line input field that you can resize at will so to fit the size of your text (two sets of position and size are remembered: one for the chat input and one for the IM input), as well as navigate in the text easily with the cursor movement keys (Left and Right, of course, but also Up and Down). Note that since the text of the input lines is automatically copied to the text input floater when it is open and since the latter is auto-focused as well, you don't risk to loose characters while the floater auto-opens (you can keep typing and everything will be fine). By pressing ENTER (or closing it manually with the "X" button), the text input floater is closed and the text is copied back into the corresponding (chat or IM) input line. This is of course a feature designed for "para-RPers" who always pester about too small a chat input line.
Made the Hide/Show Mispelled words feature a per-line/text editor setting (there is still a global setting that determines whether this feature is enabled by default on creation of each line or text editor, but disabling or enabling it in one editor doesn't affect the others any more).
Removed spell checking from the scripts editor (was not meant to use spell checking).
Added Speculoos to the default OpenSim grids list.