Cool VL Viewer v1.26.4.19 (current stable branch):
Since a few crash bugs have been fixed in this release, updating to it is recommended.

New in this release:
Exposed the "Show media HUD" option in the "Preferences" floater, "Audio & Video" tab, and fixed several media HUD bugs (including a potential crash bug related with it, in the tool pie, that could crash the viewer when hovering the mouse pointer on a bad media textures).
Implemented an "Import" feature for Shape, Hair, Skin and Eyes in the appearance customize floater. By clicking on the "Import" button in the corresponding tab, you can choose a saved archetype file (produced with the "Advanced" -> "Character" -> "Appearance to XML..." feature) so to import the various parameters (note that textures are not imported, only the "visual parameters" are). This is quite handy a feature to quickly and easily "import" your avatar's shape in other grids...
Fixed the bug that prevented changing the sex (gender) of a shape in the appearance customize floater (I wonder how this bug went unnoticed for so long... It has been introduced back in v1.26.3 !).
When voluntarily logging off from an OpenSim grid, do not ever show any more the "You have been disconnected" dialog (which is reserved for unexpected disconnections) that could spuriously appear from times to times (and even every time on some grids).
Bugfix: when a text input chat or IM floater is open, make sure that you can't send the uncompleted text from the disabled input line by focusing it and hitting ENTER.
Fixed a potential crash bug on logout, in relation with the voice service shutdown.
Better texture exporting algorithm for the Object backup feature, preventing the export action to stay stuck forever, waiting for a texture to load.
Increased the maximum size of scripts sources from 64Kb to 128Kb.
More code cleanup and (minor) speed optimizations.